feat: separar lista de produtos do formulário
